Create README.md
Hello, I'm Carlos Augusto, a recent high school graduate who is looking to improve and enter the IT and web development job market. I'm a young man who has a not extensive curriculum, but with experience in customer service at reauto, counting, in addition to experience as a mechanic. however, these other areas were discarded and today my focus is on IT and technology. I am immersed in studying in my area and also in favor of passing the ENEM in search of a vacancy in Computer Engineering. I am a young man who aspires to the right opportunity to show my real ability and to stand out in addition to leveraging and growing together with the company that gives me this opportunity. I am charismatic, educated, good listener and a great learner, who is looking for and waiting for the opportunity to learn, perfect and improve even more.